@@ -21,16 +21,21 @@
     "react-split-pane": "^0.1.92"
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.22.6",
+    "@babel/preset-react": "^7.22.5",
+    "@babel/preset-typescript": "^7.22.5",
     "@gravity-ui/eslint-config": "^1.0.2",
     "@gravity-ui/i18n": "^1.0.0",
     "@gravity-ui/prettier-config": "^1.0.1",
     "@gravity-ui/stylelint-config": "^1.0.1",
     "@gravity-ui/tsconfig": "^1.0.0",
     "@gravity-ui/uikit": "^4.2.0",
-    "@storybook/addon-essentials": "^6.5.9",
-    "@storybook/addon-knobs": "^6.4.0",
+    "@storybook/addon-essentials": "^7.0.26",
+    "@storybook/addon-knobs": "^7.0.2",
+    "@storybook/cli": "^7.0.26",
     "@storybook/preset-scss": "^1.0.3",
-    "@storybook/react": "^6.5.9",
+    "@storybook/react": "^7.0.26",
+    "@storybook/react-webpack5": "^7.0.26",
     "@types/jest": "^28.1.3",
     "@types/lodash": "^4.14.177",
     "@types/node": "^18.0.0",
@@ -58,11 +63,12 @@
     "rimraf": "^3.0.2",
     "sass": "^1.56.2",
     "sass-loader": "^10.2.1",
+    "storybook": "^7.0.26",
     "style-loader": "^2.0.0",
     "stylelint": "^14.6.0",
     "ts-jest": "^28.0.5",
     "ts-node": "^10.2.1",
-    "typescript": "^4.2.3"
+    "typescript": "^4.9.5"
   },
   "peerDependencies": {
     "@gravity-ui/uikit": "^4.0.0",
@@ -73,7 +79,7 @@
     "test": "jest",
     "test:watch": "jest --watchAll",
     "clean": "gulp clean",
-    "start": "cross-env TS_NODE_PROJECT=.storybook/tsconfig.json start-storybook -p 7007",
+    "start": "storybook dev -p 7007",
     "build": "gulp",
     "lint:js": "eslint --quiet --ext .js,.jsx,.ts,.tsx .",
     "lint:styles": "stylelint '{styles,src}/**/*.scss' --quiet",
